test: move conversion tests to a separate test file
